+++ presenter/src/main/java/org/softwaremaestro/presenter/login/CompleteStudentProfileFragment.kt
@@ -1,18 +1,16 @@
 package org.softwaremaestro.presenter.login
 
-import android.content.Intent
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels
-import androidx.navigation.Navigation
 import androidx.navigation.fragment.findNavController
 import dagger.hilt.android.AndroidEntryPoint
+import org.softwaremaestro.presenter.R
 import org.softwaremaestro.presenter.databinding.FragmentCompleteStudentProfileBinding
 import org.softwaremaestro.presenter.login.viewmodel.StudentRegisterViewModel
-import org.softwaremaestro.presenter.student_home.StudentHomeActivity
 
 @AndroidEntryPoint
 class CompleteStudentProfileFragment : Fragment() {
@@ -45,13 +43,14 @@ class CompleteStudentProfileFragment : Fragment() {
 
     private fun setRegisterResultObserver() {
         viewModel.registerSuccess.observe(viewLifecycleOwner) { isRegisterSuccess ->
+
+            // 추후에 수정하기
+            findNavController().navigate(R.id.action_completeStudentProfileFragment_to_loginFrament)
+
             if (isRegisterSuccess) {
-                val intent = Intent(requireContext(), StudentHomeActivity::class.java)
-                startActivity(intent)
             } else {
                 //TODO: 실패했을 때 처리
-                Navigation.findNavController(requireView())
-                    .popBackStack()
+
             }
         }
     }
