--- presenter/src/main/res/layout/fragment_seach_univ.xml
@@ -1,48 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".login.SearchUnivFragment">
-
-    <androidx.constraintlayout.widget.ConstraintLayout
-        android:id="@+id/container_toolbar"
-        style="@style/ToolBar"
-        tools:ignore="MissingConstraints">
-
-        <Button
-            android:id="@+id/btn_toolbar_back"
-            style="@style/ToolBarBackButton" />
-
-        <TextView
-            style="@style/ToolBarTitle"
-            android:text="질문 올리기" />
-    </androidx.constraintlayout.widget.ConstraintLayout>
-
-    <com.google.android.material.textfield.TextInputLayout
-        android:id="@+id/tf_univ"
-        style="@style/Widget.MaterialComponents.TextInputLayout.OutlinedBox"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:layout_marginHorizontal="@dimen/margin_horizontal_default"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@id/container_toolbar">
-
-        <AutoCompleteTextView
-            android:id="@+id/atv_univ"
-            style="@style/body2"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:focusable="true"
-            android:focusableInTouchMode="true"
-            android:hint="대학교"
-            android:imeOptions="actionNext"
-            android:inputType="text"
-            android:padding="16dp"
-            android:privateImeOptions="defaultInputmode=korean"
-            android:textColor="@color/black" />
-
-    </com.google.android.material.textfield.TextInputLayout>
-
-</androidx.constraintlayout.widget.ConstraintLayout>
